video_service/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pyzmq==27.1.0
websockets==15.0.1
